+++ blits.d.ts
@@ -1,0 +1,6 @@
+/// <reference types="./index.d.ts" />
+
+declare module '*.blits' {
+  const component: any;
+  export default component;
+}